requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-rest-auth==0.9.1
djangorestframework==3.6.2
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-cors-headers==2.0.2
django-rest-auth==0.9.1
djangorestframework==3.6.2
oauthlib==2.0.2
packaging==16.8
PyGithub==1.33
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-cors-headers==2.0.2
django-health-check==2.3.0
django-rest-auth==0.9.1
djangorestframework==3.6.2
oauthlib==2.0.2
packaging==16.8
PyGithub==1.33
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-cors-headers==2.0.2
django-health-check==2.3.0
django-rest-auth==0.9.1
djangorestframework==3.6.2
oauthlib==2.0.2
packaging==16.8
PyGithub==1.33
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
